requirements.txt
blinker
certifi
charset-normalizer
click
colorama
contourpy
cycler
filelock
fastapi
uvicorn[standard]
python-multipart
fonttools
fsspec
idna
itsdangerous
Jinja2
kiwisolver
MarkupSafe
matplotlib
mpmath
networkx
numpy
imageio
imageio-ffmpeg
opencv-python
packaging
pandas
pillow
psutil
py-cpuinfo
pyparsing
python-dateutil
python-dotenv
pytz
PyYAML
requests
scipy
seaborn
setuptools
six
sympy
torch
torchvision
tqdm
typing_extensions
tzdata
ultralytics
ultralytics-thop
urllib3
Werkzeug
firebase-admin
pymongo
qrcode[pil]
promptpay
paddlepaddle
paddleocr
crcmod
pyzbar
fuzzywuzzy
python-Levenshtein
gunicorn
mongoengine
pymongo[srv]
huggingface_hub
transformers
gradio_client
pyjwt
python-dateutil
uvicorn

